+++ tests/phpunit/tests/includes/collection/class-test-interactions.php
@@ -714,6 +714,7 @@ class Test_Interactions extends \WP_UnitTestCase {
 	 * Test emoji replacement in activity_to_comment.
 	 *
 	 * @covers ::activity_to_comment
+	 * @covers \Activitypub\Emoji::replace_custom_emoji
 	 */
 	public function test_activity_to_comment_with_emoji() {
 		// Mock actor with emoji in name.
